Precalculate stable/immutable expressions: executor
Cached expressions are always executed as PARAM_EXEC. Usually plans for all
cached expressions are handled during query planning and they are compiled
separatly. But sometimes they are used in dynamically loaded plans (for example,
domain constraints plans). In this case all information about them is stored in
ExprState, not in EState. Usually in this case the new node PlannedExpr is used.
It is returned by a function expression_planner, and contains an expression and
the list of its cached expressions.

Note: the cached ids (= the PARAM_EXEC ids for cached expressions) of all
non-internal cached expressions must be set by the function
set_non_internal_cachedexprs_walker after all mutations. Otherwise all the
created cached expressions are considered internal and will not be cached by
themselves.