Precalculate stable and immutable functions: planner
Now in Postgresql only immutable functions are precalculated; stable functions
are calculated for every row so in fact they don't differ from volatile
functions.

In this patch the function / operator / another expression is precalculated (=
calculated once for all output rows, but as many times as the expression is
mentioned in the query) if:
1) it doesn't return a set,
2) it's not volatile itself,
3) its arguments are also constants or precalculated expressions.

Costs are changed to reflect the changed behaviour.
Tests and small notes in the documentation are included.

Note: caching of domain constraints is not currently supported, as we cannot
assume that the planner has access to the same domain constraints that will
apply at runtime.