Allow LEFT JOINs to be removed in more cases
Normally, LEFT JOINs to base rels can only be removed if none of the
columns from the joined relation are used in the query and if the relation
has a unique index which proves the join cannot duplicate any rows on the
other side of the join.

This commit relaxes this a little to no longer require the unique index to
be present.  Alternative proofs are used in the form of DISTINCT or GROUP
BY clauses.  The idea here being that it does not matter if the join causes
row duplicate if some DISTINCT/GROUP BY will subsequently remove those
duplicate rows again.  However, there are still cases that we need to
beware of and not allow this optimization to take place.  For example, if
there are aggregate functions in the query then these will be sensitive to
these additional duplicated rows, the case is the same for windowing
functions.

On successful join removal of such a case the benefits are three-fold:

1. The join search becomes more simple.
2. No waste of effort joining the relation in the first place.
3. No waste of effort removing the additional duplicate rows caused by the
   join.