Change the executor's range table from a List to an array
When the range table contained many entries performing an rt_fetch or a
getrelid could become quite expensive due to List's O(n) lookup time.
With an array the lookup is O(1).  This improves performance quite a
bit for partitioned table UPDATEs and DELETEs which have many result
relations.

When looking for a RangeTblEntry in the executor now the new
exec_rt_fetch macro must be used. rt_fetch is no longer valid.