Implement predicate propagation for non-equivalence clauses.
Currently, planner will generate new restriction clauses that can be deduced
from equivalence clauses, with function generate_base_implied_equalities().

In this patch, we are trying to add predicate propagation for non-equivalence
clauses, that is, A=B AND f(A) implies A=B AND f(A) and f(B), under some
restrictions on f. It can be turned on/off by GUC enable_predicate_propagation.
This is ported from Greenplum database.

The idea is in distribute_qual_to_rels(), we collect a dedicated list
'non_eq_clauses' as part of PlannerInfo, which contains non-equivalence
clauses, as well as potential equivalence clauses that are rendered unsafe by
outer-join-delay considerations. For each RestrictInfo in 'non_eq_clauses', we
replace its expression with another equivalent one and make a new RestrictInfo.
For example, if a RestrictInfo stands for A>10, and if A and B are in the same
equivalence class, we generate a new RestrictInfo by replacing A with B, so we
get B>10.

This patch will introduce extra cost for relation scan, due to the cost of
evaluating the new implied quals. Meanwhile, since the extra filter may reduce
the number of tuples returned by the scan, it may lower the cost of following
joins. So, the total cost may or may not decrease, depends on the selectivity
of the implied quals. If the selectivity is low, the extra cost introduced by
the implied qual is more likely to be compromised by the cost reduced in join,
and vise versa.

Co-authored-by: Richard Guo <riguo@pivotal.io>
Co-authored-by: Alexandra Wang <lewang@pivotal.io>