Fix thinko in handling of corrupted DSM control area.
If dsm_postmaster_shutdown() determines that the DSM control segment
is corrupted, it should still destroy it, otherwise it is leaked.
Leaving it mapped in was causing an assertion failure in
dsm_postmaster_startup() during crash restart.

Reported by build-farm animal peripatus.  The root cause of the
corrupted segment in that case is not yet known, but that's a
separate problem to investigate.

Author: Thomas Munro
Reviewed-by: Amit Kapila
Diagnosed-by: Tom Lane
Discussion: https://postgr.es/m/6153.1539806400%40sss.pgh.pa.us