Allow Pathkeys to derive their order from a parent key
This parent PathKey concept has been given the name "Super Keys". The term
"super" has been borrowed out of class hierarchy from OOP. A Path
containing a PathKey which has a pk_superkey set must also be ordered by
that super key.  The reverse is not true, meaning the super key describes
a more strict ordering than any subkey which uses it.

A new column has been added to pg_proc to allow an optional setting to
mention which of the function arguments are the order key.  Giving a
function a valid order key argument is only valid for precision loss
functions or functions which return an exactly equivalent value (e.g casts
to a wider type). Some examples are:  date_trunc() truncates precision
from
the 2nd argument, so the 2nd argument can become a super PathKey allowing
a btree index on that column to provide sorted input to a GROUP BY clause
containing that function, which would allow a GroupAggregate to be used
instead of a HashAggregate, which can improve performance significantly,
especially so when only a small subset of all groups are required.

The functionality is likely also useful for casts between types, however
when modifying existing casts functions to give them an order key care
must be taken to ensure the sort order of the input and output types match
exactly.