Add costs and duration options to auto_explain
This adds additional options to auto_explain for controlling logging
of costs and the query duration. While being able to turn costs on or
off can be useful, the main driver is to be able to properly test the
extension. This commit adds basic test coverage to auto_explain using
the options introduced to keep the test output stable.