Allow lock acquisitions for partitions to be delayed
Normally during query execution, AcquireExecutorLocks will obtain locks on
all RangeTblEntry objects.  When many RangeTblEntrys are table partitions
then a large number of locks may be required.  This can slow down execution
of such plans.

Now that we have run-time partition pruning we may end up not scanning
some of these partitions and if so we have no need to obtain a lock on them.
Here we modify things so that locks are only obtained on partitions the
first time they are accessed in the executor, instead of at the start of
execution.