Reorganize collation lookup time and place
Until now, a fmgr C function that makes use of collation
information (for example varstr_cmp(), str_toupper()) gets passed the
collation OID, looks up the collation with
pg_newlocale_from_collation(), and then does something with it.

With this change, the lookup is moved earlier, typically during
executor initialization.  The fmgr functions receive the locale
pointer that is the result of that lookup.

This makes the collation lookup more similar to the function lookup.
In most cases they now happen right next to each other.  We also save
repeated lookups in repeated function calls.

Moreover, this makes the structure and logic of functions like
varstr_cmp() easier, because they don't have to deal with a mix of
OIDs and locale pointers that are sometimes not set depending on the
value of the other.