Add timeline to partial WAL segments.
If two clusters are promoted from the same timeline on the same WAL segment,
they will both archive a .partial file with the same name even if they select
different timelines to promote to.

If one partially promotes and later fails we want to be able to promote another
cluster.  This duplicate .partial file causes problems for archive commands that
won't overwrite an existing file.  Even if the archive command checks the
contents they might not be binary identical due to differing unused space at the
end of the file or because one cluster accepted more transactions.

Add the timeline being promoted *to* into the .partial filename to prevent the
conflict, e.g. 000000010000000100000001.00000002.partial.

If more than one cluster tries to promote to the same timeline there will still
be a conflict but this is desirable.  Once a cluster has claimed a timeline no
other cluster should be archiving on it.