Allow user to use password instead of encryption key.
pg_keytool reads the password from stdin, uses "key derivation function (KDF)
parameters" stored in the data directory to derive the key and writes the key
to standard output.

The cluster can be created this way

	initdb -D data -K "echo securepwd | pg_keytool -D data  -w"

and started either this way

	pg_ctl -D data -K "echo securepwd | pg_keytool -D data  -w" start

or this way

	postgres -D data

and (in another session)

	echo securepwd | pg_keytool -D data -ws

Both initdb and pg_ctl can substitute the data directory for %D pattern in the
key encryption command. Thus user uses pg_keytool to derive the password, he
can in fact say

	initdb -D data -K "echo securepwd | pg_keytool -D %D  -w"

and

	pg_ctl -D data -K "echo securepwd | pg_keytool -D %D  -w" start

respectively.