Make buffile.c aware of encryption.
This includes temporary files as well as the transient files introduced by
this patch series.

One particular problem encountered here is that the whole multiple of
encryption block needs to be written, backend needs to know how many bytes at
the end of file segment is just padding. The backend can easily remember it
for normal temporary file, but for shared temporary file also this information
needs to be shared. Currently we append 4-byte word at the end of the segment
which informs on length of the useful data. Should this information be shared
via shared memory instead?

No padding would be needed if we used stream cipher, but that's not safe,
README.encryption)