Introduce "transient buffered file".
Currently we need it only to encrypt transaction data which gets spilled to
disk during logical decoding. This way we hide the encryption specific details
from reorderbuffer.c, as well as any other (future) user of transient files
that need to be encrypted.

The next patch of the series implements the actual encryption of those files.