Add tests for buffile.c.
Since the encryption adds complexity to the buffile.c, regression tests seem
like a good idea. In order to be able to test corner cases that deal with
segment boudaries, this patch makes the segment size configurable - the tests
would be too expensive with 1GB segments.

The new GUC is marked with DEVELOPER_OPTIONS, DBA is not supposed to use it
during normal operation.