Add EXPLAIN option jit_details showing per-expression information about JIT.
This is useful both to understand where JIT is applied (and thus where
to improve), and to be able write regression tests to verify that we
can JIT compile specific parts of a query.

Note that currently the printed function names will make it harder to
use this for regression tests - a followup commit will improve that
angle.

Author:
Reviewed-By:
Discussion: https://postgr.es/m/
Backpatch: