Refactoring SPI execute related errors.
Deprecating SPI_ERROR_PARAM in favor of throwing an error.  This error
code was only being returned when a caller of SPI_execute_plan,
SPI_execute_snapshot, or SPI_execute_with_args passed in a plan which
has nargs > 0 but then negligently passed in NULL for the Values and/or
argtypes.  This would be a C coding error, not a state of affairs that a
user could trigger.