Move AtEOXact_Snapshot() back to AbortTransaction().
As a general rule, it seems preferable to do as much work as possible
in AbortTransaction(), leaving as little work as possible to be done
in CleanupTransaction(), because it's a good idea to hold on to
resources for the shortest possible amount of time.

Way back in 2011, commit 57eb009092684e6e1788dd0dae641ccee1668b10
moved AbortTransaction's AtEOXact_Snapshot call to
CleanupTransaction to fix a problem when a ROLLBACK statement was
prepared at the protocol level and executed in a transaction
with REPEATABLE READ or higher isolation. RevalidateCachedQuery
would attempt to obtain a snapshot and fail because the transaction
was already gone.  Commit ac63dca607e8e22247defbc8fe03b6baa3628c42
subsequently taught RevalidateCachedQuery not to obtain a snapshot
for such commands after all while fixing an unrelated bug, and
there now seems to be no case in which we obtain a snapshot in
an aborted transaction.

That's good, because it can't possibly be safe to use a snapshot in
an aborted transaction, for reasons explained in the comments
added by this patch. So, move the call back to AbortTransaction().
To help catch future cases of unsafe snapshot use, add a bunch of
Assert() statements to snapmgr.c to flag dangerous coding patterns.
Flesh out the comments in RevalidateCachedQuery a bit, too.

Patch by me, reviewed by Andres Freund.

Discussion: http://postgr.es/m/CA+TgmoZKo449=6cTzRkVvaWVpOJT__Cwy=wFtW32bB31sNRLfg@mail.gmail.com