Remove single chararter literal tokens from the core grammar
By default, when Bison generates token symbols, it starts numbering them
at 258, so that numbers below 256 can be used for character literals. The
internal token number is looked up via the yytranslate[] array. Bison 3.5
has a new directive "%define api.token.raw", which causes Bison to write
out the same ("raw") token numbers it would use internally, and thus skip
building the yytranslate[] array as well as the code to access it.

To be compatible with this directive, arrange for the core scanner to map
single characters to new named tokens. Any invalid single characters now
cause an error in the scanner rather than the parser.

The speedup in raw parsing (Flex + Bison) is modest, at best 2%, and it'll
be years before we can make the new directive the default, but making it
possible to remove one useless array access and its cache pollution seems
a worthwhile thing to do.