Teach MemoryContext infrastructure not to depend on Node.
Previously, each type of MemoryContext was a separate type of
node. Now, there's a single magic number for all types, and the
specific type of context is indicated only by the "methods"
pointer.

This has a couple of advantages. First, it means that extensions are
free to create new memory context tyeps.  Second, it means that
nodes/memnodes.h no longer needs to depend on nodes/nodes.h, which
could be useful if we someday want to allow the use of this
infrastructure in frontend code (and reducing dependencies is not a
bad thing, anyway).

Along the way, convert MemoryContextIsValid() from a macro to an
inline function, as per newer practice, and thereby remove the
multiple evaluation hazard.

Discussion: http://postgr.es/m/CAOP8fzbsfUtka3Lo02guuAWEKFTej6AcsEVQsLmYmBgoMOuRWA@mail.gmail.com