Doc update for the server message not being displayed when the client connection is closed by the server.
Doc update for the server message not being displayed when the client
connection is closed by the server in case of
idle_in_transaction_session_timeout, pg_terminate_backend, pg_ctl kill TERM
and drop database with (force). When the server closes client connection, it
sends the ErrorResponse packet, and then closes the socket and terminates the
backend process. If the packet is received before the server closes the
connection, the error message is received in both windows and linux. If the
packet is not received before the server closes the connection, the error
message is not received in case of windows where as in non-windows it is
received.