Allow ALTER TABLE to do an index scan, like CLUSTER
The idea is to implement table rewrite in ALTER more similar to what CLUSTER
does, to allow clustering during table-rewriting ALTER.

Note, this does not do the AM-specific visibility checks that CLUSTER does, so
will not clean up dead tuples.

The indices are normally dropped before doing table rewrite, so index scan
would be impossible.  As a POC, this early implementation avoids dropping the
indices.