Use centralized StartSubprocess for aux procs
Modifies StartChildProcess and moves the auxiliary processes under its
control. This is the next step toward making use of the Subprocess
infrastructure.

This patch introduces the fork_failure function, which may be
implemented by each subprocess to control failure when attempting to
fork the subprocess. This is used by the startup subprocess to exit
postmaster on fork failure. Future subprocesses may use it to do cleanup
(and optionally exit postmaster) after a fork failure.

Additionally, the patch adds a new backend variable for establishing a
backend variable to be shared across fork/exec.