Build root tuple in ExecBuildSlotValueDescription
Having the logic to do so be spread across various callers is
starting to seem cumbersome to maintain.

Use mt_root_tuple_slot to store the tuple to be shown in the
description. For it to be available in all cases that require it,
always set it in ExecInitModifyTable instead if only when UPDATE
tuple routing is active as is the case currently.