Add equalCompare() function to compare node types
Traditionally we've done this with equal() which just returns true if the
nodes are equal and false otherwise.  equalCompare() provides a comparison
result for node types so that we can do things like sort the nodes and
binary search for specific node values.  The equal() function has now just
become a simple wrapper around equalCompare() and returns true if the
comparison value is 0.

No actual usages of the new function, besides the call made by equal()
have been added here.  This just adds the core infrastructure to allow
further work to be done later.

For extension authors whose extensions make use of an extended node type,
the ExtensibleNodeMethods.nodeEqual() function now no longer exists. This
must be replaced with nodeCompare(), which follows in the footsteps of
equalCompare() and returns < 0 when a < b, > 0 when a > b and 0 when a is
equal to b.