Implementing the cmdstats subsystem.
This implements a subsystem for tracking the number of times a type
of command has been run in a database cluster, since startup or
since the last time the counts were reset, whichever is newer.

This work is based on and borrows from work by Haribabu Kommi
(haribabu) as reflected in the patch submission:

  SQL statements statistics counter view (pg_stat_sql)

which itself was reviewed by dilip kumar (dilip.kumar), Vinayak
Pokale (vinpokale) and probably others.