Support leading minus in front of ISO 8601 interval string
It is a common extension for a standard to have negative intervals
prefixed with a leading minus instead of individual parts.
For example -PT1H as an alternative to PT-1H.
The standard itself is not specific on negative intervals.

This patch makes PostgreSQL a bit more compatible with 3-rd party software
that use ISO 8601 style intervals that might be negative, e.g.
* https://github.com/rails/rails/issues/39503
* https://docs.oracle.com/javase/8/docs/api/java/time/Duration.html#parse-java.lang.CharSequence-