Process multiple RI trigger events at a time.
It should be more efficient to execute the check query once per multiple rows
inserted / updated / deleted than to run the query for every single row again.

If the user query only affects a single row, the simple query is still used to
check the RI, as it was before this patch.