Introduce infrastructure for batch processing RI events.
Separate storage is used for the RI trigger events because the "transient
table" that we provide to statement triggers would not be available for
deferred constraints. Also, the regular statement level trigger is not ideal
for the RI checks because it requires the query execution to complete before
the RI checks even start. On the other hand, if we use batches of row trigger
events, we only need to tune the batch size so that user gets RI violation
error rather soon.

This patch only introduces the infrastructure, however the trigger function is
still called per event. This is just to reduce the size of the diffs.