Provide a GUC to allow vacuum to continue on corrupted tuple
A new GUC to control whether the vacuum will error out immediately
on detection of a corrupted tuple or it will just emit a WARNING for
each such instance and complete the rest of the vacuuming.  However,
nothing will be done to the corrupted tuple itself and we will also
not advance the relfrozenxid and relminmxid.