Detect the null type correctly in setop case by peeking the later exprs.
In the past, we will get error when issue select NULL union select NULL
union select 1;  that is mainly because we set the expr type to Text after
handling the first 2 expr with some reason. Now before we do that, we
first peak the first not null expr to get the real type we can use, so we can
overcome this issue.