Fix GUC parse_int() to allow fractional input only when a unit is accepted.
The following commit made some changes to how integer GUC option values are
parsed and what values are allowed, allowing fractional input and rounding
to the nearest multiple of the next smaller unit:

https://github.com/postgres/postgres/commit/1a83a80a2fe5b559f85ed4830acb92d5124b7a9a

However, it also allowed fractional input for cases it probably shouldn't
have (i.e. when the setting does not accept a unit), such as:

log_file_mode = 384.234
max_connections = 1.0067e2
port = 5432.123

Also, parse_int() is used for parsing other options, such as the integer
storage parameters for CREATE TABLE and CREATE INDEX. For example, the
following are currently allowed but probably shouldn't be:

CREATE TABLE ... WITH (fillfactor = 23.45);
CREATE TABLE ... WITH (parallel_workers = 5.4);

A couple of related test items are affected by this fix.